+++ example/lib/nested.g.dart
@@ -7,8 +7,11 @@ part of 'nested.dart';
 // **************************************************************************
 
 // ignore_for_file: duplicate_ignore
-// ignore_for_file: unused_local_variable
+// ignore_for_file: omit_local_variable_types
+// ignore_for_file: prefer_function_declarations_over_variables
+// ignore_for_file: unnecessary_brace_in_string_interps
 // ignore_for_file: unnecessary_string_interpolations
+// ignore_for_file: unused_local_variable
 @GenerateFor(_myTemplate)
 TrustedHtml myTemplate() {
   var $ = StringBuffer();
@@ -25,8 +28,11 @@ TrustedHtml myTemplate() {
 }
 
 // ignore_for_file: duplicate_ignore
-// ignore_for_file: unused_local_variable
+// ignore_for_file: omit_local_variable_types
+// ignore_for_file: prefer_function_declarations_over_variables
+// ignore_for_file: unnecessary_brace_in_string_interps
 // ignore_for_file: unnecessary_string_interpolations
+// ignore_for_file: unused_local_variable
 @GenerateFor(_img)
 TrustedHtml img(String url) {
   var $ = StringBuffer();