+++ example/test/template_test.g.dart
@@ -7,8 +7,11 @@ part of 'template_test.dart';
 // **************************************************************************
 
 // ignore_for_file: duplicate_ignore
-// ignore_for_file: unused_local_variable
+// ignore_for_file: omit_local_variable_types
+// ignore_for_file: prefer_function_declarations_over_variables
+// ignore_for_file: unnecessary_brace_in_string_interps
 // ignore_for_file: unnecessary_string_interpolations
+// ignore_for_file: unused_local_variable
 @GenerateFor(_emptyTemplate)
 TrustedHtml emptyTemplate() {
   var $ = StringBuffer();
@@ -17,8 +20,11 @@ TrustedHtml emptyTemplate() {
 }
 
 // ignore_for_file: duplicate_ignore
-// ignore_for_file: unused_local_variable
+// ignore_for_file: omit_local_variable_types
+// ignore_for_file: prefer_function_declarations_over_variables
+// ignore_for_file: unnecessary_brace_in_string_interps
 // ignore_for_file: unnecessary_string_interpolations
+// ignore_for_file: unused_local_variable
 @GenerateFor(_simpleTemplate)
 TrustedHtml simpleTemplate() {
   var $ = StringBuffer();
@@ -29,8 +35,11 @@ TrustedHtml simpleTemplate() {
 }
 
 // ignore_for_file: duplicate_ignore
-// ignore_for_file: unused_local_variable
+// ignore_for_file: omit_local_variable_types
+// ignore_for_file: prefer_function_declarations_over_variables
+// ignore_for_file: unnecessary_brace_in_string_interps
 // ignore_for_file: unnecessary_string_interpolations
+// ignore_for_file: unused_local_variable
 @GenerateFor(_condition)
 TrustedHtml condition({bool? myVar = false}) {
   var $ = StringBuffer();
@@ -43,8 +52,11 @@ TrustedHtml condition({bool? myVar = false}) {
 }
 
 // ignore_for_file: duplicate_ignore
-// ignore_for_file: unused_local_variable
+// ignore_for_file: omit_local_variable_types
+// ignore_for_file: prefer_function_declarations_over_variables
+// ignore_for_file: unnecessary_brace_in_string_interps
 // ignore_for_file: unnecessary_string_interpolations
+// ignore_for_file: unused_local_variable
 @GenerateFor(_interpolation)
 TrustedHtml interpolation(String? someText) {
   var $ = StringBuffer();
@@ -55,8 +67,11 @@ TrustedHtml interpolation(String? someText) {
 }
 
 // ignore_for_file: duplicate_ignore
-// ignore_for_file: unused_local_variable
+// ignore_for_file: omit_local_variable_types
+// ignore_for_file: prefer_function_declarations_over_variables
+// ignore_for_file: unnecessary_brace_in_string_interps
 // ignore_for_file: unnecessary_string_interpolations
+// ignore_for_file: unused_local_variable
 @GenerateFor(_loop)
 TrustedHtml loop(Iterable<String>? list) {
   var $ = StringBuffer();
@@ -72,8 +87,11 @@ TrustedHtml loop(Iterable<String>? list) {
 }
 
 // ignore_for_file: duplicate_ignore
-// ignore_for_file: unused_local_variable
+// ignore_for_file: omit_local_variable_types
+// ignore_for_file: prefer_function_declarations_over_variables
+// ignore_for_file: unnecessary_brace_in_string_interps
 // ignore_for_file: unnecessary_string_interpolations
+// ignore_for_file: unused_local_variable
 @GenerateFor(_switchTemplate)
 TrustedHtml switchTemplate(MyEnum? myEnum) {
   var $ = StringBuffer();
@@ -105,8 +123,11 @@ TrustedHtml switchTemplate(MyEnum? myEnum) {
 }
 
 // ignore_for_file: duplicate_ignore
-// ignore_for_file: unused_local_variable
+// ignore_for_file: omit_local_variable_types
+// ignore_for_file: prefer_function_declarations_over_variables
+// ignore_for_file: unnecessary_brace_in_string_interps
 // ignore_for_file: unnecessary_string_interpolations
+// ignore_for_file: unused_local_variable
 @GenerateFor(_attribute)
 TrustedHtml attribute({
   bool disabled = false,